pyproject.toml
[tool.poetry]
name = "joatss"
version = "0.1.0"
description = ""
authors = ["이규락 <62545703+leegyurak@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
anthropic = "^0.34.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
